Refactor training code
