+++ train_model.py
@@ -1,72 +1,16 @@
-import os
 import time
 
 import matplotlib.pyplot as plt
 import numpy as np
-import pandas as pd
 import torch
 import torch.nn as nn
 import torch.optim as optim
-from torch.utils.data import DataLoader, Dataset, random_split
+from torch.utils.data import DataLoader, random_split
 from tqdm import tqdm
 from sklearn.metrics import mean_absolute_error, r2_score
 
-
-# Define dataset class
-class VehicleDataset(Dataset):
-    def __init__(self, data_dir, sequence_length=30):
-        self.sequence_length = sequence_length
-        self.data = self.load_data(data_dir)
-
-    def load_data(self, data_dir):
-        all_files = [os.path.join(data_dir, f) for f in os.listdir(data_dir) if f.endswith('.csv')]
-        dataframes = [pd.read_csv(f) for f in all_files]
-        data = pd.concat(dataframes, ignore_index=True)
-
-        # Strip column names of any leading/trailing whitespace
-        data.columns = data.columns.str.strip()
-
-        # Strip spaces from all values and replace non-numeric values with NaN
-        data = data.map(lambda x: str(x).strip() if isinstance(x, str) else x)
-
-        # Select relevant input features and target variables
-        input_features = [
-            "Altitude", "Bearing", "Air Fuel Ratio(Measured)(:1)",
-            "Engine Load(%)", "Engine Load(Absolute)(%)",
-            "Engine RPM(rpm)", "Intake Air Temperature(°F)", "Relative Throttle Position(%)",
-            "Speed (OBD)(mph)"
-        ]
-        target_features = ["Miles Per Gallon(Instant)(mpg)", "Fuel used (inst)"]
-        all_features = input_features + target_features
-        data = data[all_features]
-        data.replace('-', np.nan, inplace=True)
-        data = data.apply(pd.to_numeric, errors='coerce')
-        data = data.dropna()
-
-        X = data[input_features].astype(np.float32).values
-        y = data[target_features].astype(np.float32).values
-        return X, y
-
-    def __len__(self):
-        return len(self.data[0]) - self.sequence_length
-
-    def __getitem__(self, idx):
-        X_seq = self.data[0][idx:idx + self.sequence_length]
-        y_seq = self.data[1][idx + self.sequence_length - 1]  # Predict the last time step
-        return torch.tensor(X_seq), torch.tensor(y_seq)
-
-
-# Define LSTM model
-class FuelMPGRNN(nn.Module):
-    def __init__(self, input_size, hidden_size, num_layers, output_size):
-        super(FuelMPGRNN, self).__init__()
-        self.lstm = nn.LSTM(input_size, hidden_size, num_layers, batch_first=True)
-        self.fc = nn.Linear(hidden_size, output_size)
-
-    def forward(self, x):
-        lstm_out, _ = self.lstm(x)
-        out = self.fc(lstm_out[:, -1, :])
-        return out
+from mpg_rnn.fuel_mpg_rnn import FuelMPGRNN
+from data_loading.vehicle_dataset import VehicleDataset
 
 
 # Load dataset
@@ -81,8 +25,8 @@ dataloader_train = DataLoader(train_dataset, batch_size=128, shuffle=True)
 dataloader_val = DataLoader(val_dataset, batch_size=128, shuffle=False)
 
 # Define model parameters
-input_size = 9  # Number of input features
-hidden_size = 64
+input_size = 8  # Number of input features
+hidden_size = 16
 num_layers = 2
 output_size = 2  # Predicting 2 variables
 
@@ -101,6 +45,7 @@ optimizer = optim.Adam(model.parameters(), lr=0.001)
 # Training loop
 epochs = 100
 gpu_time = 0
+max_r2 = float("-inf")
 for epoch in range(epochs):
     start = time.perf_counter()
     model.train()
@@ -152,11 +97,13 @@ for epoch in range(epochs):
     r2_train = r2_score(train_targets, train_preds)
 
     print(f"Epoch {epoch + 1}/{epochs}, Loss: {loss_total:.4f}, Val Loss: {val_loss:.4f}, RMSE: {rmse:.4f}, "
-          f"MAE: {mae:.4f}, R²: {r2:.4f}, R² train: {r2_train:.4f}")
+          f"MAE: {mae:.4f}, R²: {r2:.4f}, R² train: {r2_train:.4f}, R² max: {max_r2:.4f}")
     gpu_time += end - start
     # For TQDM
     time.sleep(0.01)
-    # torch.save(model, "checkpoint.pth")
+    if r2 > max_r2:
+        max_r2 = r2
+        torch.save(model, "checkpoint.pth")
 
 print("Training complete.")
 print(f"GPU time: {gpu_time:.4f}s")
